adk-agent/requirements.txt
google-adk>=0.1.0
google-genai>=0.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
